requirements.txt
# requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
pandas>=1.3.0
pyyaml>=6.0
tqdm>=4.62.0
open3d>=0.15.0  # For point cloud visualization and processing
numba>=0.55.0  # For performance optimization
jupyter>=1.0.0
notebook>=6.4.0
seaborn>=0.11.0
